+++ src/features/order/MenuList.js
@@ -1,8 +1,7 @@
 import styled from '@emotion/styled';
 import React, {useState} from 'react';
 import { useRecoilValue } from 'recoil';
-import { MenuListState, FilterListState } from '../../store';
-import {Button} from '../../components';
+import { MenuListState } from '../../store';
 
 // TODO : 메뉴 HOT, ICE 선택 - Input Change
 // TODO : 초기화
@@ -16,62 +15,13 @@ import {Button} from '../../components';
 // MEMO : 주문 정보 => store에 존재 해야함, 취소 버튼을 누르기 전까지 주문 정보 유지
 
 export function MenuList() {
-  const filterConditionList = useRecoilValue(FilterListState);
   const menuList = useRecoilValue(MenuListState);
-  const numRegex = /[0-9]/g;
-  const handleSearch = () => {
-    let searchResult = [];
-    if(numRegex.test(searchValue)) {
-      menuList.items.forEach((item, index) => {
-        const size = item.productPrice.size;
-        const sizeKeys = Object.keys(size);
-        sizeKeys.forEach((sizeKey) => {
-          if(size[sizeKey] === +searchValue) searchResult.push(item);
-        });
-      });
-    } else {
-      searchResult = menuList.items.filter((item) => item.productName.indexOf(searchValue) > -1);
-    }
-  };
-  const handleInit = () => {
-    console.log('초기화');
-  }
-  const handleFilter = (type, filterValue) => (e) => {
-    const ENG_REGEX =/[a-zA-Z]/g;
-
-    let filterResult = [];
-    filterResult = menuList.items.filter((value) => {
-      switch (type) {
-        case 'size' :
-          const sizeKeys = Object.keys(value.productPrice.size);
-          sizeKeys.includes(filterValue)
-          return sizeKeys.includes(filterValue);
-        case 'isOnlyIce' :
-          return value.isOnlyIce;
-        default :
-          if(ENG_REGEX.test(filterValue)) {
-            const valueUpperCase = filterValue.toUpperCase();
-            const keyLabel = Object.keys(value.ingredientLabel).map((item) => item.toUpperCase());
-            return keyLabel.includes(valueUpperCase);
-          } else {
-            const keyInfo = Object.keys(value.ingredientLabel).map((item) => value.ingredientLabel[item].ko);
-            return keyInfo.includes(filterValue);
-          }
-      }
-    });
-    // 상품 목록 진하게 표시
-  };
   const handleOptionChange = (itemName) => e => {
     if(e.target.checked) {
       setOptionValue(itemName);
       // 주문 진행
     }
   }
-  const handleSearchChange = (e) => {
-    setSearchValue(e.target.value);
-  }
-
-  const [searchValue, setSearchValue] = useState('');
   const [optionValue, setOptionValue] = useState('');
 
   return (
@@ -82,22 +32,6 @@ export function MenuList() {
       리코일 사용법을 위해 간단한 예시를 둠
       분할이 필요하면 분할 할 것
        */}
-      <strong>메뉴 검색</strong>
-      <input type="text" id="searchKeyword" title="검색" value={searchValue} onChange={handleSearchChange}/>
-      <Button onClick={handleSearch} text="검색"/>
-      <strong>필터 LIST</strong>
-      <ul>
-        {
-          filterConditionList.map((item) => {
-            return (
-              <li key={`item_${item.id}`}>
-                <Button onClick={handleFilter(item.type, item.text)} text={item.text} />
-              </li>
-            );
-          })
-        }
-      </ul>
-      <Button onClick={handleInit} text="초기화"/>
       <strong>MENU LIST</strong>
       <ul>
         {
